Mark consumer groups as not supported in tests
I am under the impression that consumer groups are not actually supported. I believe that the statement "consumer groups are supported" actually means something along the lines of "multiple subscribers listening to the same subscription will not be fighting over messages and a message will be delivered to only one of them". This statement seems to be false.